test(session): 拆分会话管理大文件测试
