+++ src-tauri/src/session_management_tests.rs
@@ -1,181 +1,3 @@
-    use super::*;
-    use crate::types::{WorkspaceKind, WorkspaceSettings, WorktreeInfo};
-    use std::io::Write;
-    use uuid::Uuid;
-
-    fn workspace_entry(
-        id: &str,
-        name: &str,
-        path: &str,
-        kind: WorkspaceKind,
-        parent_id: Option<&str>,
-    ) -> WorkspaceEntry {
-        WorkspaceEntry {
-            id: id.to_string(),
-            name: name.to_string(),
-            path: path.to_string(),
-            codex_bin: None,
-            kind: kind.clone(),
-            parent_id: parent_id.map(ToString::to_string),
-            worktree: if kind.is_worktree() {
-                Some(WorktreeInfo {
-                    branch: "feature/test".to_string(),
-                    base_ref: None,
-                    base_commit: None,
-                    tracking: None,
-                    publish_error: None,
-                    publish_retry_command: None,
-                })
-            } else {
-                None
-            },
-            settings: WorkspaceSettings::default(),
-        }
-    }
-
-    fn catalog_entry(
-        session_id: &str,
-        workspace_id: &str,
-        workspace_label: Option<&str>,
-        cwd: Option<&str>,
-    ) -> WorkspaceSessionCatalogEntry {
-        WorkspaceSessionCatalogEntry {
-            session_id: session_id.to_string(),
-            stable_session_key: None,
-            canonical_session_id: Some(session_id.to_string()),
-            parent_session_id: None,
-            workspace_id: workspace_id.to_string(),
-            workspace_label: workspace_label.map(ToString::to_string),
-            engine: "codex".to_string(),
-            title: "Example session".to_string(),
-            updated_at: 1,
-            archived_at: None,
-            thread_kind: "native".to_string(),
-            source: Some("cli".to_string()),
-            source_label: Some("cli/codex".to_string()),
-            provider_profile_id: None,
-            provider_profile_source: None,
-            provider_profile_name: None,
-            provider_availability: None,
-            source_completeness: None,
-            source_status_reason: None,
-            size_bytes: None,
-            cwd: cwd.map(ToString::to_string),
-            attribution_status: None,
-            attribution_reason: None,
-            attribution_confidence: None,
-            matched_workspace_id: None,
-            matched_workspace_label: None,
-            folder_id: None,
-            auto_session: None,
-            exists_on_disk: true,
-            inconsistency_code: None,
-            delete_mode: Some(SESSION_DELETE_MODE_PHYSICAL.to_string()),
-            physical_path: None,
-            children_count: None,
-        }
-    }
-
-    fn write_codex_session_fixture(
-        codex_home: &Path,
-        session_id: &str,
-        cwd: &str,
-    ) -> std::path::PathBuf {
-        write_codex_session_fixture_with_message(
-            codex_home,
-            session_id,
-            cwd,
-            "2026-01-19T12:00:00.000Z",
-            "2026-01-19T12:00:05.000Z",
-            "Fixture session",
-        )
-    }
-
-    fn write_codex_session_fixture_with_message(
-        codex_home: &Path,
-        session_id: &str,
-        cwd: &str,
-        metadata_timestamp: &str,
-        message_timestamp: &str,
-        message: &str,
-    ) -> std::path::PathBuf {
-        let day_dir = codex_home
-            .join("sessions")
-            .join("2026")
-            .join("01")
-            .join("19");
-        std::fs::create_dir_all(&day_dir).expect("create codex fixture day dir");
-        let path = day_dir.join(format!("{session_id}.jsonl"));
-        let mut file = std::fs::File::create(&path).expect("create codex fixture");
-        writeln!(
-                file,
-                r#"{{"timestamp":"{metadata_timestamp}","type":"session_meta","payload":{{"id":"{session_id}","cwd":"{cwd}"}}}}"#
-            )
-            .expect("write codex fixture metadata");
-        writeln!(
-                file,
-                r#"{{"timestamp":"{message_timestamp}","type":"response_item","payload":{{"type":"message","role":"user","content":[{{"type":"input_text","text":"{message}"}}]}}}}"#
-            )
-            .expect("write codex fixture message");
-        path
-    }
-
-    fn create_claude_project_dir(base_dir: &Path, workspace_path: &Path) -> std::path::PathBuf {
-        let encoded = workspace_path
-            .to_string_lossy()
-            .chars()
-            .map(|c| {
-                if c.is_ascii_alphanumeric() || c == '-' {
-                    c
-                } else {
-                    '-'
-                }
-            })
-            .collect::<String>();
-        let project_dir = base_dir.join(encoded);
-        std::fs::create_dir_all(&project_dir).expect("create claude project dir");
-        project_dir
-    }
-
-    fn write_claude_session_fixture(
-        claude_projects_dir: &Path,
-        workspace_path: &Path,
-        session_id: &str,
-        cwd: &Path,
-        message: &str,
-    ) {
-        let project_dir = create_claude_project_dir(claude_projects_dir, workspace_path);
-        let session_path = project_dir.join(format!("{session_id}.jsonl"));
-        let mut file = std::fs::File::create(session_path).expect("create claude fixture");
-        writeln!(
-                file,
-                r#"{{"uuid":"user-1","timestamp":"2026-01-19T12:00:00.000Z","session_id":"{session_id}","cwd":"{}","message":{{"role":"user","content":"{message}"}}}}"#,
-                cwd.to_string_lossy()
-            )
-            .expect("write claude fixture");
-    }
-
-    fn codex_fixture_timestamp(minutes_before_latest: usize) -> String {
-        let latest_total_minutes: usize = 20 * 60;
-        let total_minutes = latest_total_minutes.saturating_sub(minutes_before_latest);
-        format!(
-            "2026-01-19T{:02}:{:02}:00.000Z",
-            total_minutes / 60,
-            total_minutes % 60
-        )
-    }
-
-    fn workspace_with_codex_home(
-        id: &str,
-        name: &str,
-        path: &str,
-        codex_home: &Path,
-    ) -> WorkspaceEntry {
-        let mut workspace = workspace_entry(id, name, path, WorkspaceKind::Main, None);
-        workspace.settings.codex_home = Some(codex_home.to_string_lossy().to_string());
-        workspace
-    }
-
     #[test]
     fn parses_prefixed_and_stable_cursor() {
         assert_eq!(parse_catalog_cursor(Some("offset:25")), 25);
@@ -205,7 +27,6 @@
             SessionCatalogCursor::LegacyOffset(_) => panic!("expected stable cursor"),
         }
     }
-
     #[test]
     fn catalog_scan_limit_uses_requested_page_window_plus_lookahead() {
         assert_eq!(build_catalog_scan_limit(None, Some(25)), 26);
@@ -588,2586 +409,4 @@
             .folder_id_by_session_id
             .get("claude:child:session-1")
             .is_none());
-    }
-
-    #[test]
-    fn metadata_orphan_entries_wait_for_authoritative_engine_source() {
-        let workspace = workspace_entry("ws-1", "Workspace", "/repo", WorkspaceKind::Main, None);
-        let metadata = WorkspaceSessionCatalogMetadata {
-            archived_at_by_session_id: HashMap::from([
-                ("claude:ws-1:session-1".to_string(), 42),
-                ("codex:ws-1:session-2".to_string(), 43),
-            ]),
-            ..Default::default()
-        };
-        let mut incomplete_entries = Vec::new();
-
-        append_metadata_orphan_entries(
-            &mut incomplete_entries,
-            &workspace,
-            &metadata,
-            &[
-                WorkspaceSessionCatalogSourceStatus {
-                    engine: "claude".to_string(),
-                    source_kind: None,
-            completeness: WorkspaceSessionSourceCompleteness::UncertainEmpty,
-                    reason: Some(SESSION_CATALOG_PARTIAL_CLAUDE_UNCERTAIN_EMPTY.to_string()),
-                    scanned_candidates: Some(0),
-                    skipped_candidates: None,
-                    scan_cap_reached: Some(false),
-                    diagnostics: Vec::new(),
-                    cache: None,
-                },
-                WorkspaceSessionCatalogSourceStatus {
-                    engine: "codex".to_string(),
-                    source_kind: None,
-            completeness: WorkspaceSessionSourceCompleteness::Complete,
-                    reason: None,
-                    scanned_candidates: Some(0),
-                    skipped_candidates: None,
-                    scan_cap_reached: Some(false),
-                    diagnostics: Vec::new(),
-                    cache: None,
-                },
-            ],
-        );
-
-        assert!(!incomplete_entries
-            .iter()
-            .any(|entry| entry.engine == "claude"));
-        assert!(incomplete_entries
-            .iter()
-            .any(|entry| entry.engine == "codex"));
-
-        let mut authoritative_entries = Vec::new();
-        append_metadata_orphan_entries(
-            &mut authoritative_entries,
-            &workspace,
-            &metadata,
-            &[WorkspaceSessionCatalogSourceStatus {
-                engine: "claude".to_string(),
-                source_kind: None,
-            completeness: WorkspaceSessionSourceCompleteness::AuthoritativeEmpty,
-                reason: None,
-                scanned_candidates: Some(0),
-                skipped_candidates: None,
-                scan_cap_reached: Some(false),
-                diagnostics: Vec::new(),
-                cache: None,
-            }],
-        );
-
-        assert!(authoritative_entries
-            .iter()
-            .any(|entry| entry.engine == "claude"
-                && entry.delete_mode.as_deref() == Some(SESSION_DELETE_MODE_METADATA_CLEANUP)));
-    }
-
-    #[tokio::test]
-    async fn auto_session_metadata_persists_and_projects_visibility() {
-        let base = std::env::temp_dir().join(format!("auto-session-metadata-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-
-        record_auto_session_metadata_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-hidden".to_string(),
-            AutoSessionMetadata {
-                session_purpose: "prompt-enhancer".to_string(),
-                visibility: AutoSessionVisibility::Hidden,
-                owner_feature: "composer".to_string(),
-                auto_archive: Some(true),
-                created_by: AutoSessionCreatedBy::System,
-            },
-        )
-        .await
-        .expect("record hidden metadata");
-        record_auto_session_metadata_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-traceable".to_string(),
-            AutoSessionMetadata {
-                session_purpose: "spec-hub-apply".to_string(),
-                visibility: AutoSessionVisibility::SystemAuto,
-                owner_feature: "spec-hub".to_string(),
-                auto_archive: Some(false),
-                created_by: AutoSessionCreatedBy::System,
-            },
-        )
-        .await
-        .expect("record system-auto metadata");
-
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        let metadata_by_workspace_id = HashMap::from([("ws-1".to_string(), metadata)]);
-        let hidden = finalize_existing_catalog_entry(
-            catalog_entry("codex-hidden", "ws-1", Some("Workspace"), None),
-            &metadata_by_workspace_id,
-        );
-        let traceable = finalize_existing_catalog_entry(
-            catalog_entry("codex-traceable", "ws-1", Some("Workspace"), None),
-            &metadata_by_workspace_id,
-        );
-
-        assert!(entry_is_hidden_automatic_session(&hidden));
-        assert_eq!(
-            traceable.folder_id.as_deref(),
-            Some(SESSION_FOLDER_SYSTEM_AUTO_ID)
-        );
-        assert_eq!(
-            traceable
-                .auto_session
-                .as_ref()
-                .map(|metadata| metadata.session_purpose.as_str()),
-            Some("spec-hub-apply")
-        );
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[test]
-    fn provider_home_catalog_entry_projects_unavailable_provider_without_disk_fallback() {
-        let mut entry = catalog_entry("codex-provider-session", "ws-1", Some("Workspace"), None);
-        entry.provider_profile_id = Some("deleted-provider".to_string());
-        entry.provider_profile_source = Some("managed".to_string());
-        entry.provider_availability = Some("unknown".to_string());
-
-        let finalized = finalize_existing_catalog_entry(entry, &HashMap::new());
-
-        assert_eq!(
-            finalized.provider_profile_id.as_deref(),
-            Some("deleted-provider")
-        );
-        assert_eq!(finalized.provider_profile_source.as_deref(), Some("managed"));
-        assert_eq!(
-            finalized.provider_profile_name.as_deref(),
-            Some("deleted-provider")
-        );
-        assert_eq!(
-            finalized.provider_availability.as_deref(),
-            Some("unavailable")
-        );
-        assert_eq!(finalized.source_label.as_deref(), Some("deleted-provider"));
-    }
-
-    #[test]
-    fn mutation_target_resolves_provider_home_codex_row() {
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = HashMap::from([(workspace.id.clone(), workspace)]);
-        let mut entry = catalog_entry("provider-session", "ws-1", Some("Workspace"), None);
-        entry.provider_profile_id = Some("provider-a".to_string());
-        entry.provider_profile_source = Some("managed".to_string());
-        entry.provider_profile_name = Some("Provider A".to_string());
-        entry.provider_availability = Some("available".to_string());
-        entry.physical_path = Some(
-            "/tmp/app/codex-provider-homes/provider-a/sessions/2026/01/19/provider-session.jsonl"
-                .to_string(),
-        );
-
-        let target = resolve_session_mutation_target(&[entry], &workspaces, "provider-session")
-            .expect("provider-home row should resolve as mutation target");
-
-        assert_eq!(target.owner_workspace_id, "ws-1");
-        assert_eq!(target.native_session_id, "provider-session");
-        assert_eq!(target.provider_profile_id.as_deref(), Some("provider-a"));
-        assert!(target.exists_on_disk);
-        assert_eq!(
-            target.delete_mode.as_deref(),
-            Some(SESSION_DELETE_MODE_PHYSICAL)
-        );
-    }
-
-    #[tokio::test]
-    async fn provider_home_only_codex_session_restores_from_catalog_without_runtime() {
-        let base = std::env::temp_dir().join(format!("provider-home-catalog-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let workspace_path = base.join("workspace");
-        std::fs::create_dir_all(&workspace_path).expect("create workspace path");
-        let provider_home = base.join("codex-provider-homes").join("provider-a");
-        write_codex_session_fixture(
-            &provider_home,
-            "provider-home-only",
-            &workspace_path.to_string_lossy(),
-        );
-
-        let mut workspace = workspace_entry(
-            "ws-1",
-            "Workspace",
-            &workspace_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        workspace.settings.codex_home = Some(provider_home.to_string_lossy().to_string());
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-
-        let page = list_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            Some(WorkspaceSessionCatalogQuery {
-                engine: Some("codex".to_string()),
-                status: Some("all".to_string()),
-                ..Default::default()
-            }),
-            None,
-            Some(20),
-        )
-        .await
-        .expect("list provider-home catalog sessions");
-
-        let entry = page
-            .data
-            .iter()
-            .find(|entry| entry.session_id == "provider-home-only")
-            .expect("provider-home session should be restored from catalog");
-        assert_eq!(entry.provider_profile_id.as_deref(), Some("provider-a"));
-        assert_eq!(entry.provider_profile_source.as_deref(), Some("managed"));
-        assert_eq!(entry.provider_profile_name.as_deref(), Some("provider-a"));
-        assert_eq!(entry.provider_availability.as_deref(), Some("unavailable"));
-        assert!(entry.exists_on_disk);
-        assert!(page.source_statuses.iter().any(|status| {
-            status.engine == "codex" && status.source_kind.as_deref() == Some("disk")
-        }));
-        assert!(page.source_statuses.iter().any(|status| {
-            status.engine == "codex" && status.source_kind.as_deref() == Some("provider-home")
-        }));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn provider_home_archive_folder_and_delete_target_only_session_file() {
-        let base = std::env::temp_dir().join(format!("provider-home-mutate-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let workspace_path = base.join("workspace");
-        std::fs::create_dir_all(&workspace_path).expect("create workspace path");
-        let provider_home = base.join("codex-provider-homes").join("provider-a");
-        let target_path = write_codex_session_fixture(
-            &provider_home,
-            "provider-delete-target",
-            &workspace_path.to_string_lossy(),
-        );
-        let sibling_path = write_codex_session_fixture(
-            &provider_home,
-            "provider-keep-sibling",
-            &workspace_path.to_string_lossy(),
-        );
-
-        let mut workspace = workspace_entry(
-            "ws-1",
-            "Workspace",
-            &workspace_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        workspace.settings.codex_home = Some(provider_home.to_string_lossy().to_string());
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-        let requested_session_id = "provider-delete-target".to_string();
-        let stable_key = "codex:ws-1:provider-delete-target".to_string();
-
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Provider folder".to_string(),
-            None,
-        )
-        .await
-        .expect("create provider folder")
-        .folder;
-        let folder_response = assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            requested_session_id.clone(),
-            Some(folder.id.clone()),
-        )
-        .await
-        .expect("assign provider-home folder");
-        assert_eq!(folder_response.folder_id.as_deref(), Some(folder.id.as_str()));
-
-        let archive_response = archive_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            vec![requested_session_id.clone()],
-        )
-        .await
-        .expect("archive provider-home session");
-        assert!(archive_response.results[0].ok);
-
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert_eq!(
-            metadata.folder_id_by_session_id.get(&stable_key),
-            Some(&folder.id)
-        );
-        assert!(metadata.archived_at_by_session_id.contains_key(&stable_key));
-
-        let delete_response = delete_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            vec![requested_session_id],
-        )
-        .await
-        .expect("delete provider-home session");
-        assert!(delete_response.results[0].ok);
-        assert_eq!(delete_response.results[0].deleted_from_disk, Some(true));
-        assert!(!target_path.exists());
-        assert!(sibling_path.exists());
-        assert!(provider_home.exists());
-
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read cleaned metadata");
-        assert!(!metadata.folder_id_by_session_id.contains_key(&stable_key));
-        assert!(!metadata.archived_at_by_session_id.contains_key(&stable_key));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn provider_home_mutation_failure_reports_provider_aware_diagnostic() {
-        let base =
-            std::env::temp_dir().join(format!("provider-home-unresolved-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let workspace_path = base.join("workspace");
-        std::fs::create_dir_all(&workspace_path).expect("create workspace path");
-        let mut workspace = workspace_entry(
-            "ws-1",
-            "Workspace",
-            &workspace_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        workspace.settings.codex_home = Some(
-            base.join("codex-provider-homes")
-                .join("provider-a")
-                .to_string_lossy()
-                .to_string(),
-        );
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-
-        let response = delete_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            vec!["codex:missing-provider-session".to_string()],
-        )
-        .await
-        .expect("unresolved provider-home delete should return batch result");
-
-        assert!(!response.results[0].ok);
-        assert_eq!(
-            response.results[0].code.as_deref(),
-            Some("OWNER_WORKSPACE_UNRESOLVED")
-        );
-        assert!(response.results[0]
-            .error
-            .as_deref()
-            .unwrap_or_default()
-            .contains("provider-home source"));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn auto_session_metadata_rejects_invalid_boundary_values() {
-        let base = std::env::temp_dir().join(format!("auto-session-invalid-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-
-        let valid_metadata = AutoSessionMetadata {
-            session_purpose: "prompt-enhancer".to_string(),
-            visibility: AutoSessionVisibility::Hidden,
-            owner_feature: "composer".to_string(),
-            auto_archive: Some(true),
-            created_by: AutoSessionCreatedBy::System,
-        };
-
-        let invalid_session = record_auto_session_metadata_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "../escape".to_string(),
-            valid_metadata.clone(),
-        )
-        .await
-        .expect_err("invalid session id rejected");
-        assert_eq!(invalid_session, "invalid session_id");
-
-        let empty_purpose = record_auto_session_metadata_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-valid".to_string(),
-            AutoSessionMetadata {
-                session_purpose: "   ".to_string(),
-                ..valid_metadata.clone()
-            },
-        )
-        .await
-        .expect_err("empty session purpose rejected");
-        assert_eq!(empty_purpose, "sessionPurpose is required");
-
-        let path_like_owner = record_auto_session_metadata_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-valid".to_string(),
-            AutoSessionMetadata {
-                owner_feature: "feature\\nested".to_string(),
-                ..valid_metadata
-            },
-        )
-        .await
-        .expect_err("path-like owner feature rejected");
-        assert_eq!(path_like_owner, "invalid ownerFeature");
-
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(metadata.auto_session_by_session_id.is_empty());
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn system_auto_metadata_exposes_reserved_folder_group() {
-        let base = std::env::temp_dir().join(format!("system-auto-folder-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-
-        let empty_tree = list_workspace_session_folders_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-        )
-        .await
-        .expect("list empty folders");
-        assert!(
-            empty_tree
-                .folders
-                .iter()
-                .all(|folder| folder.id != SESSION_FOLDER_SYSTEM_AUTO_ID)
-        );
-
-        record_auto_session_metadata_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-traceable".to_string(),
-            AutoSessionMetadata {
-                session_purpose: "spec-hub-apply".to_string(),
-                visibility: AutoSessionVisibility::SystemAuto,
-                owner_feature: "spec-hub".to_string(),
-                auto_archive: Some(false),
-                created_by: AutoSessionCreatedBy::System,
-            },
-        )
-        .await
-        .expect("record system-auto metadata");
-
-        let tree = list_workspace_session_folders_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-        )
-        .await
-        .expect("list system-auto folders");
-        let system_folder = tree
-            .folders
-            .iter()
-            .find(|folder| folder.id == SESSION_FOLDER_SYSTEM_AUTO_ID)
-            .expect("system-auto folder");
-        assert_eq!(system_folder.workspace_id, "ws-1");
-        assert_eq!(system_folder.parent_id, None);
-        assert_eq!(system_folder.name, "system-auto");
-
-        assert!(create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Invalid".to_string(),
-            Some(SESSION_FOLDER_SYSTEM_AUTO_ID.to_string()),
-        )
-        .await
-        .is_err());
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_session_folder_tree_starts_empty_and_persists_nested_folders() {
-        let base = std::env::temp_dir().join(format!("session-folders-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-
-        let empty = list_workspace_session_folders_core(&workspaces, &storage_path, "ws-1".to_string())
-            .await
-            .expect("list empty tree");
-        assert_eq!(empty.folders, Vec::<WorkspaceSessionFolder>::new());
-
-        let parent = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Bugs".to_string(),
-            None,
-        )
-        .await
-        .expect("create parent folder")
-        .folder;
-        let child = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Regression".to_string(),
-            Some(parent.id.clone()),
-        )
-        .await
-        .expect("create child folder")
-        .folder;
-
-        let tree = list_workspace_session_folders_core(&workspaces, &storage_path, "ws-1".to_string())
-            .await
-            .expect("list populated tree");
-        assert_eq!(tree.folders.len(), 2);
-        assert!(tree.folders.iter().any(|folder| folder.id == parent.id));
-        assert!(tree
-            .folders
-            .iter()
-            .any(|folder| folder.id == child.id
-                && folder.parent_id.as_deref() == Some(parent.id.as_str())));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_session_folder_crud_rejects_cycles_and_deletes_empty_subtree() {
-        let base = std::env::temp_dir().join(format!("session-folder-crud-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-
-        let parent = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Parent".to_string(),
-            None,
-        )
-        .await
-        .expect("create parent")
-        .folder;
-        let child = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Child".to_string(),
-            Some(parent.id.clone()),
-        )
-        .await
-        .expect("create child")
-        .folder;
-
-        let cycle_error = move_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            parent.id.clone(),
-            Some(child.id.clone()),
-        )
-        .await
-        .expect_err("cycle move must fail");
-        assert_eq!(cycle_error, "folder tree cannot contain cycles");
-
-        delete_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            parent.id.clone(),
-        )
-        .await
-        .expect("delete empty folder subtree");
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(!metadata.folders.iter().any(|folder| folder.id == parent.id));
-        assert!(!metadata.folders.iter().any(|folder| folder.id == child.id));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_session_folder_delete_cleans_stale_subtree_assignments_when_visible_count_is_zero(
-    ) {
-        let base = std::env::temp_dir().join(format!("session-folder-stale-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Stale folder".to_string(),
-            None,
-        )
-        .await
-        .expect("create folder")
-        .folder;
-        let child = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Stale child".to_string(),
-            Some(folder.id.clone()),
-        )
-        .await
-        .expect("create child folder")
-        .folder;
-        with_catalog_metadata_mutation(&storage_path, "ws-1", |metadata| {
-            metadata
-                .folder_id_by_session_id
-                .insert("codex:ws-1:missing-session".to_string(), child.id.clone());
-            Ok(())
-        })
-        .expect("write stale folder assignment");
-
-        delete_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            folder.id.clone(),
-        )
-        .await
-        .expect("delete empty folder with stale metadata");
-
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(!metadata.folders.iter().any(|item| item.id == folder.id));
-        assert!(!metadata.folders.iter().any(|item| item.id == child.id));
-        assert!(metadata.folder_id_by_session_id.is_empty());
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_session_folder_delete_promotes_existing_session_assignment_to_root() {
-        let base = std::env::temp_dir().join(format!("session-folder-real-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        write_codex_session_fixture(&codex_home, "codex-real", "/tmp/ws-1");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Real folder".to_string(),
-            None,
-        )
-        .await
-        .expect("create folder")
-        .folder;
-        assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-real".to_string(),
-            Some(folder.id.clone()),
-        )
-        .await
-        .expect("assign real session");
-
-        delete_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            folder.id.clone(),
-        )
-        .await
-        .expect("delete folder with real session");
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(!metadata.folders.iter().any(|item| item.id == folder.id));
-        assert!(!metadata
-            .folder_id_by_session_id
-            .contains_key(&metadata_stable_key_for_session_id("ws-1", "codex-real")));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_session_folder_delete_promotes_child_session_assignment_to_parent_folder() {
-        let base = std::env::temp_dir().join(format!("session-folder-child-real-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        write_codex_session_fixture(&codex_home, "codex-child-real", "/tmp/ws-1");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-
-        let parent = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Parent".to_string(),
-            None,
-        )
-        .await
-        .expect("create parent folder")
-        .folder;
-        let child = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Child".to_string(),
-            Some(parent.id.clone()),
-        )
-        .await
-        .expect("create child folder")
-        .folder;
-        assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-child-real".to_string(),
-            Some(child.id.clone()),
-        )
-        .await
-        .expect("assign real session to child folder");
-
-        delete_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            child.id.clone(),
-        )
-        .await
-        .expect("delete child folder with real session");
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(metadata.folders.iter().any(|item| item.id == parent.id));
-        assert!(!metadata.folders.iter().any(|item| item.id == child.id));
-        assert_eq!(
-            metadata
-                .folder_id_by_session_id
-                .get(&metadata_stable_key_for_session_id(
-                    "ws-1",
-                    "codex-child-real"
-                )),
-            Some(&parent.id)
-        );
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_session_folder_delete_promotes_descendant_session_assignment_to_root() {
-        let base = std::env::temp_dir().join(format!("session-folder-subtree-real-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        write_codex_session_fixture(&codex_home, "codex-descendant-real", "/tmp/ws-1");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-
-        let parent = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Parent".to_string(),
-            None,
-        )
-        .await
-        .expect("create parent folder")
-        .folder;
-        let child = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Child".to_string(),
-            Some(parent.id.clone()),
-        )
-        .await
-        .expect("create child folder")
-        .folder;
-        assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-descendant-real".to_string(),
-            Some(child.id.clone()),
-        )
-        .await
-        .expect("assign real session to child folder");
-
-        delete_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            parent.id.clone(),
-        )
-        .await
-        .expect("delete parent folder subtree with real session");
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(!metadata.folders.iter().any(|item| item.id == parent.id));
-        assert!(!metadata.folders.iter().any(|item| item.id == child.id));
-        assert!(!metadata
-            .folder_id_by_session_id
-            .contains_key(&metadata_stable_key_for_session_id(
-                "ws-1",
-                "codex-descendant-real"
-            )));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn session_folder_assignment_supports_same_workspace_and_root_fallback() {
-        let base = std::env::temp_dir().join(format!("session-folder-assign-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        write_codex_session_fixture(&codex_home, "codex-1", "/tmp/ws-1");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Bugs".to_string(),
-            None,
-        )
-        .await
-        .expect("create folder")
-        .folder;
-
-        let assigned = assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-1".to_string(),
-            Some(folder.id.clone()),
-        )
-        .await
-        .expect("assign session");
-        assert_eq!(assigned.folder_id.as_deref(), Some(folder.id.as_str()));
-
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert_eq!(
-            metadata
-                .folder_id_by_session_id
-                .get(&metadata_stable_key_for_session_id("ws-1", "codex-1")),
-            Some(&folder.id)
-        );
-
-        let root = assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-1".to_string(),
-            Some(SESSION_FOLDER_ROOT_ID.to_string()),
-        )
-        .await
-        .expect("move to root");
-        assert_eq!(root.folder_id, None);
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(!metadata.folder_id_by_session_id.contains_key("codex-1"));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn session_folder_assignment_accepts_session_beyond_first_owner_lookup_page() {
-        let base = std::env::temp_dir().join(format!("session-folder-deep-owner-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        for index in 0..=205 {
-            let timestamp = codex_fixture_timestamp(index);
-            write_codex_session_fixture_with_message(
-                &codex_home,
-                &format!("codex-{index:03}"),
-                "/tmp/ws-1",
-                &timestamp,
-                &timestamp,
-                "Fixture session",
-            );
-        }
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Deep".to_string(),
-            None,
-        )
-        .await
-        .expect("create folder")
-        .folder;
-
-        let assigned = assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-205".to_string(),
-            Some(folder.id.clone()),
-        )
-        .await
-        .expect("deep session still belongs to workspace");
-
-        assert_eq!(assigned.folder_id.as_deref(), Some(folder.id.as_str()));
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[test]
-    fn folder_assignment_is_applied_to_catalog_entries_by_owner_workspace() {
-        let mut entry = catalog_entry("claude:1", "ws-1", Some("Workspace"), None);
-        let metadata_by_workspace_id = HashMap::from([(
-            "ws-1".to_string(),
-            WorkspaceSessionCatalogMetadata {
-                folder_id_by_session_id: HashMap::from([(
-                    "claude:1".to_string(),
-                    "folder-1".to_string(),
-                )]),
-                ..Default::default()
-            },
-        )]);
-
-        apply_folder_assignment(&mut entry, &metadata_by_workspace_id);
-
-        assert_eq!(entry.folder_id.as_deref(), Some("folder-1"));
-    }
-
-    #[test]
-    fn codex_folder_assignment_accepts_raw_and_prefixed_session_keys() {
-        let mut raw_entry = catalog_entry("codex-1", "ws-1", Some("Workspace"), None);
-        raw_entry.engine = "codex".to_string();
-        let mut prefixed_entry = catalog_entry("codex:codex-2", "ws-1", Some("Workspace"), None);
-        prefixed_entry.engine = "codex".to_string();
-        let metadata_by_workspace_id = HashMap::from([(
-            "ws-1".to_string(),
-            WorkspaceSessionCatalogMetadata {
-                folder_id_by_session_id: HashMap::from([
-                    ("codex:codex-1".to_string(), "folder-1".to_string()),
-                    ("codex-2".to_string(), "folder-2".to_string()),
-                ]),
-                ..Default::default()
-            },
-        )]);
-
-        apply_folder_assignment(&mut raw_entry, &metadata_by_workspace_id);
-        apply_folder_assignment(&mut prefixed_entry, &metadata_by_workspace_id);
-
-        assert_eq!(raw_entry.folder_id.as_deref(), Some("folder-1"));
-        assert_eq!(prefixed_entry.folder_id.as_deref(), Some("folder-2"));
-    }
-
-    #[test]
-    fn codex_folder_assignment_cleanup_removes_raw_and_prefixed_keys() {
-        let mut metadata = WorkspaceSessionCatalogMetadata {
-            folder_id_by_session_id: HashMap::from([
-                ("codex-1".to_string(), "folder-raw".to_string()),
-                ("codex:codex-1".to_string(), "folder-prefixed".to_string()),
-                ("claude:1".to_string(), "folder-claude".to_string()),
-            ]),
-            ..Default::default()
-        };
-
-        remove_folder_assignment_for_session(&mut metadata, "ws-1", "codex-1", "codex");
-
-        assert!(!metadata.folder_id_by_session_id.contains_key("codex-1"));
-        assert!(!metadata
-            .folder_id_by_session_id
-            .contains_key("codex:codex-1"));
-        assert!(metadata.folder_id_by_session_id.contains_key("claude:1"));
-    }
-
-    #[tokio::test]
-    async fn archive_evidence_reads_metadata_without_catalog_scan_and_expands_stable_keys() {
-        let base = std::env::temp_dir().join(format!("archive-evidence-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        with_catalog_metadata_mutation(&storage_path, "ws-1", |metadata| {
-            metadata
-                .archived_at_by_session_id
-                .insert("claude:ws-1:session-1".to_string(), 123);
-            metadata
-                .archived_at_by_session_id
-                .insert("codex:ws-1:codex-1".to_string(), 456);
-            Ok(())
-        })
-        .expect("seed archive metadata");
-
-        let evidence = list_workspace_session_archive_evidence_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-        )
-        .await
-        .expect("read archive evidence");
-
-        assert_eq!(
-            evidence.archived_at_by_session_id.get("claude:session-1"),
-            Some(&123)
-        );
-        assert_eq!(
-            evidence
-                .archived_at_by_session_id
-                .get("claude:ws-1:session-1"),
-            Some(&123)
-        );
-        assert_eq!(
-            evidence.archived_at_by_session_id.get("codex-1"),
-            Some(&456)
-        );
-        assert_eq!(
-            evidence.archived_at_by_session_id.get("codex:codex-1"),
-            Some(&456)
-        );
-        assert_eq!(evidence.partial_source, None);
-        assert_eq!(
-            evidence.source_statuses[0].completeness,
-            WorkspaceSessionSourceCompleteness::Complete
-        );
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn archive_evidence_reports_partial_when_metadata_unavailable() {
-        let base = std::env::temp_dir().join(format!("archive-evidence-bad-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let metadata_path = catalog_metadata_path(&storage_path, "ws-1").expect("metadata path");
-        std::fs::create_dir_all(metadata_path.parent().expect("metadata parent"))
-            .expect("create metadata parent");
-        std::fs::write(&metadata_path, "{not-json").expect("write corrupt metadata");
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-
-        let evidence = list_workspace_session_archive_evidence_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-        )
-        .await
-        .expect("read degraded archive evidence");
-
-        assert!(evidence.archived_at_by_session_id.is_empty());
-        assert_eq!(
-            evidence.partial_source.as_deref(),
-            Some(SESSION_CATALOG_PARTIAL_ARCHIVE_METADATA)
-        );
-        assert_eq!(evidence.source_statuses.len(), 1);
-        assert_eq!(evidence.source_statuses[0].engine, "archive-metadata");
-        assert_eq!(
-            evidence.source_statuses[0].completeness,
-            WorkspaceSessionSourceCompleteness::Degraded
-        );
-        assert_eq!(
-            evidence.source_statuses[0].reason.as_deref(),
-            Some(SESSION_CATALOG_PARTIAL_ARCHIVE_METADATA)
-        );
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[test]
-    fn folder_sorting_is_deterministic_by_name_created_at_and_id() {
-        let mut folders = vec![
-            WorkspaceSessionFolder {
-                id: "b".to_string(),
-                workspace_id: "ws-1".to_string(),
-                parent_id: None,
-                name: "Zeta".to_string(),
-                created_at: 1,
-                updated_at: 1,
-            },
-            WorkspaceSessionFolder {
-                id: "c".to_string(),
-                workspace_id: "ws-1".to_string(),
-                parent_id: None,
-                name: "Alpha".to_string(),
-                created_at: 2,
-                updated_at: 2,
-            },
-            WorkspaceSessionFolder {
-                id: "a".to_string(),
-                workspace_id: "ws-1".to_string(),
-                parent_id: None,
-                name: "Alpha".to_string(),
-                created_at: 1,
-                updated_at: 1,
-            },
-        ];
-
-        sort_workspace_session_folders(&mut folders);
-
-        let ids: Vec<_> = folders.into_iter().map(|folder| folder.id).collect();
-        assert_eq!(ids, vec!["a", "c", "b"]);
-    }
-
-    #[tokio::test]
-    async fn session_folder_assignment_rejects_missing_target_without_rewriting_previous_assignment() {
-        let base = std::env::temp_dir().join(format!("session-folder-missing-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        write_codex_session_fixture(&codex_home, "codex-1", "/tmp/ws-1");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Bugs".to_string(),
-            None,
-        )
-        .await
-        .expect("create folder")
-        .folder;
-        assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-1".to_string(),
-            Some(folder.id.clone()),
-        )
-        .await
-        .expect("assign session");
-
-        let error = assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-1".to_string(),
-            Some("missing-folder".to_string()),
-        )
-        .await
-        .expect_err("missing folder must fail");
-        assert_eq!(error, "target folder not found");
-
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert_eq!(
-            metadata
-                .folder_id_by_session_id
-                .get(&metadata_stable_key_for_session_id("ws-1", "codex-1")),
-            Some(&folder.id)
-        );
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn session_folder_assignment_rejects_folder_from_other_workspace() {
-        let base = std::env::temp_dir().join(format!("session-folder-cross-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home_1 = base.join("codex-home-1");
-        let codex_home_2 = base.join("codex-home-2");
-        write_codex_session_fixture(&codex_home_1, "codex-1", "/tmp/ws-1");
-        let ws_1 = workspace_with_codex_home("ws-1", "Workspace 1", "/tmp/ws-1", &codex_home_1);
-        let ws_2 = workspace_with_codex_home("ws-2", "Workspace 2", "/tmp/ws-2", &codex_home_2);
-        let workspaces = Mutex::new(HashMap::from([
-            (ws_1.id.clone(), ws_1),
-            (ws_2.id.clone(), ws_2),
-        ]));
-        let engine_manager = engine::EngineManager::new();
-        let other_folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-2".to_string(),
-            "Other".to_string(),
-            None,
-        )
-        .await
-        .expect("create other workspace folder")
-        .folder;
-
-        let error = assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-1".to_string(),
-            Some(other_folder.id),
-        )
-        .await
-        .expect_err("cross-workspace folder assignment must fail");
-
-        assert_eq!(error, "target folder not found");
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn session_folder_assignment_rejects_wrong_project_and_preserves_metadata() {
-        let base = std::env::temp_dir().join(format!("session-folder-owner-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home_1 = base.join("codex-home-1");
-        let codex_home_2 = base.join("codex-home-2");
-        write_codex_session_fixture(&codex_home_2, "codex-other", "/tmp/ws-2");
-        let ws_1 = workspace_with_codex_home("ws-1", "Workspace 1", "/tmp/ws-1", &codex_home_1);
-        let ws_2 = workspace_with_codex_home("ws-2", "Workspace 2", "/tmp/ws-2", &codex_home_2);
-        let workspaces = Mutex::new(HashMap::from([
-            (ws_1.id.clone(), ws_1),
-            (ws_2.id.clone(), ws_2),
-        ]));
-        let engine_manager = engine::EngineManager::new();
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Target".to_string(),
-            None,
-        )
-        .await
-        .expect("create target folder")
-        .folder;
-        let preserved = WorkspaceSessionCatalogMetadata {
-            archived_at_by_session_id: HashMap::from([("codex-keep".to_string(), 42)]),
-            ..Default::default()
-        };
-        with_catalog_metadata_mutation(&storage_path, "ws-1", |metadata| {
-            *metadata = preserved;
-            Ok(())
-        })
-        .expect("seed metadata");
-
-        let error = assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-other".to_string(),
-            Some(folder.id),
-        )
-        .await
-        .expect_err("wrong-project session must fail");
-
-        assert_eq!(error, "session does not belong to target workspace");
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert_eq!(
-            metadata
-                .archived_at_by_session_id
-                .get("codex-keep")
-                .copied(),
-            Some(42)
-        );
-        assert!(!metadata.folder_id_by_session_id.contains_key("codex-other"));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn session_folder_assignment_rejects_unresolved_session_owner_without_writing() {
-        let base = std::env::temp_dir().join(format!("session-folder-unresolved-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Target".to_string(),
-            None,
-        )
-        .await
-        .expect("create target folder")
-        .folder;
-
-        let error = assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-missing".to_string(),
-            Some(folder.id),
-        )
-        .await
-        .expect_err("unresolved session must fail");
-
-        assert_eq!(error, "session does not belong to target workspace");
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(!metadata
-            .folder_id_by_session_id
-            .contains_key("codex-missing"));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn archive_preserves_folder_assignment_and_active_filter_hides_session() {
-        let base = std::env::temp_dir().join(format!("session-archive-folder-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        write_codex_session_fixture(&codex_home, "codex-keep", "/tmp/ws-1");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Keep".to_string(),
-            None,
-        )
-        .await
-        .expect("create folder")
-        .folder;
-        assign_workspace_session_folder_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            "codex-keep".to_string(),
-            Some(folder.id.clone()),
-        )
-        .await
-        .expect("assign session");
-
-        archive_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            vec!["codex-keep".to_string()],
-        )
-        .await
-        .expect("archive session");
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert_eq!(
-            metadata
-                .folder_id_by_session_id
-                .get(&metadata_stable_key_for_session_id("ws-1", "codex-keep")),
-            Some(&folder.id)
-        );
-
-        let entry = WorkspaceSessionCatalogEntry {
-            archived_at: metadata
-                .archived_at_by_session_id
-                .get(&metadata_stable_key_for_session_id("ws-1", "codex-keep"))
-                .copied(),
-            ..catalog_entry("codex-keep", "ws-1", Some("Workspace"), None)
-        };
-        assert!(!entry_matches_status(
-            &entry,
-            SessionCatalogStatusFilter::Active
-        ));
-        assert!(entry_matches_status(
-            &entry,
-            SessionCatalogStatusFilter::Archived
-        ));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_session_list_keyword_finds_match_beyond_first_scan_window() {
-        let base = std::env::temp_dir().join(format!("session-keyword-deep-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        for index in 0..60 {
-            let timestamp = codex_fixture_timestamp(index);
-            let message = if index == 55 {
-                "Needle regression"
-            } else {
-                "Ordinary session"
-            };
-            write_codex_session_fixture_with_message(
-                &codex_home,
-                &format!("codex-{index:03}"),
-                "/tmp/ws-1",
-                &timestamp,
-                &timestamp,
-                message,
-            );
-        }
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-
-        let page = list_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            Some(WorkspaceSessionCatalogQuery {
-                keyword: Some("needle".to_string()),
-                engine: None,
-                status: Some("all".to_string()),
-                folder_id: None,
-                ..Default::default()
-            }),
-            None,
-            Some(10),
-        )
-        .await
-        .expect("list sessions");
-
-        assert_eq!(page.data.len(), 1);
-        assert_eq!(page.data[0].session_id, "codex-055");
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn projection_summary_counts_full_history_beyond_default_scan_window() {
-        let base = std::env::temp_dir().join(format!("session-summary-full-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        for index in 0..60 {
-            let timestamp = codex_fixture_timestamp(index);
-            write_codex_session_fixture_with_message(
-                &codex_home,
-                &format!("codex-{index:03}"),
-                "/tmp/ws-1",
-                &timestamp,
-                &timestamp,
-                "Fixture session",
-            );
-        }
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let engine_manager = engine::EngineManager::new();
-
-        let summary = get_workspace_session_projection_summary_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            Some(WorkspaceSessionCatalogQuery {
-                keyword: None,
-                engine: None,
-                status: Some("all".to_string()),
-                folder_id: None,
-                ..Default::default()
-            }),
-        )
-        .await
-        .expect("summary");
-
-        assert_eq!(summary.all_total, 60);
-        assert_eq!(summary.filtered_total, 60);
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[test]
-    fn delete_success_metadata_cleanup_removes_global_and_folder_state() {
-        let mut metadata = WorkspaceSessionCatalogMetadata {
-            archived_at_by_session_id: HashMap::from([("claude:gone".to_string(), 42_i64)]),
-            folder_id_by_session_id: HashMap::from([(
-                "claude:gone".to_string(),
-                "folder-a".to_string(),
-            )]),
-            ..Default::default()
-        };
-
-        metadata.archived_at_by_session_id.remove("claude:gone");
-        metadata.folder_id_by_session_id.remove("claude:gone");
-
-        assert!(!metadata
-            .archived_at_by_session_id
-            .contains_key("claude:gone"));
-        assert!(!metadata.folder_id_by_session_id.contains_key("claude:gone"));
-    }
-
-    #[test]
-    fn keyword_match_includes_source_fields() {
-        let entry = catalog_entry("codex:abc", "ws-1", None, None);
-
-        assert!(entry_matches_keyword(&entry, "example"));
-        assert!(entry_matches_keyword(&entry, "codex"));
-        assert!(entry_matches_keyword(&entry, "cli/codex"));
-    }
-
-    #[test]
-    fn missing_delete_errors_are_treated_as_settled_success() {
-        assert!(should_settle_delete_as_success(
-            "[SESSION_NOT_FOUND] Session file not found: stale-session"
-        ));
-        assert!(should_settle_delete_as_success(
-            "thread not found: stale-thread"
-        ));
-        assert!(!should_settle_delete_as_success(
-            "[SESSION_NOT_FOUND] Invalid OpenCode session id"
-        ));
-        assert!(!should_settle_delete_as_success("permission denied"));
-        assert!(!should_settle_delete_as_success("workspace not connected"));
-    }
-
-    #[tokio::test]
-    async fn orphan_metadata_is_listed_for_cleanup_in_all_scope() {
-        let base = std::env::temp_dir().join(format!("session-orphan-list-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let workspace = workspace_entry("ws-1", "Workspace", "/tmp/ws-1", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Stale".to_string(),
-            None,
-        )
-        .await
-        .expect("create folder")
-        .folder;
-        with_catalog_metadata_mutation(&storage_path, "ws-1", |metadata| {
-            metadata
-                .archived_at_by_session_id
-                .insert("codex-missing".to_string(), 42);
-            metadata
-                .folder_id_by_session_id
-                .insert("codex-missing".to_string(), folder.id.clone());
-            Ok(())
-        })
-        .expect("write orphan metadata");
-
-        let page = list_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            Some(WorkspaceSessionCatalogQuery {
-                keyword: None,
-                engine: None,
-                status: Some("all".to_string()),
-                folder_id: None,
-                ..Default::default()
-            }),
-            None,
-            Some(20),
-        )
-        .await
-        .expect("list sessions");
-
-        let orphan = page
-            .data
-            .iter()
-            .find(|entry| entry.session_id == "codex-missing")
-            .expect("orphan entry");
-        assert!(!orphan.exists_on_disk);
-        assert_eq!(
-            orphan.inconsistency_code.as_deref(),
-            Some(SESSION_INCONSISTENCY_MISSING_ON_DISK)
-        );
-        assert_eq!(
-            orphan.delete_mode.as_deref(),
-            Some(SESSION_DELETE_MODE_METADATA_CLEANUP)
-        );
-        assert_eq!(orphan.folder_id.as_deref(), Some(folder.id.as_str()));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn delete_missing_session_cleans_orphan_metadata_successfully() {
-        let base = std::env::temp_dir().join(format!("session-orphan-delete-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-        let codex_home = base.join("codex-home");
-        let workspace = workspace_with_codex_home("ws-1", "Workspace", "/tmp/ws-1", &codex_home);
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "ws-1".to_string(),
-            "Stale".to_string(),
-            None,
-        )
-        .await
-        .expect("create folder")
-        .folder;
-        with_catalog_metadata_mutation(&storage_path, "ws-1", |metadata| {
-            metadata
-                .archived_at_by_session_id
-                .insert("codex-missing".to_string(), 42);
-            metadata
-                .folder_id_by_session_id
-                .insert("codex-missing".to_string(), folder.id.clone());
-            Ok(())
-        })
-        .expect("write orphan metadata");
-
-        let response = delete_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "ws-1".to_string(),
-            vec!["codex-missing".to_string()],
-        )
-        .await
-        .expect("delete missing session");
-
-        assert_eq!(response.results.len(), 1);
-        assert!(response.results[0].ok);
-        assert_eq!(
-            response.results[0].code.as_deref(),
-            Some(SESSION_DELETE_CODE_ALREADY_MISSING_CLEANED)
-        );
-        assert_eq!(response.results[0].deleted_from_disk, Some(false));
-        assert_eq!(response.results[0].metadata_cleaned, Some(true));
-        let metadata = read_catalog_metadata(&storage_path, "ws-1").expect("read metadata");
-        assert!(!metadata
-            .archived_at_by_session_id
-            .contains_key("codex-missing"));
-        assert!(!metadata
-            .folder_id_by_session_id
-            .contains_key("codex-missing"));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn catalog_workspace_scope_includes_child_worktrees_for_main_workspace() {
-        let main = workspace_entry("main", "Main", "/tmp/main", WorkspaceKind::Main, None);
-        let worktree_b = workspace_entry(
-            "worktree-b",
-            "B",
-            "/tmp/worktree-b",
-            WorkspaceKind::Worktree,
-            Some("main"),
-        );
-        let worktree_a = workspace_entry(
-            "worktree-a",
-            "A",
-            "/tmp/worktree-a",
-            WorkspaceKind::Worktree,
-            Some("main"),
-        );
-        let unrelated = workspace_entry("other", "Other", "/tmp/other", WorkspaceKind::Main, None);
-        let workspaces = Mutex::new(HashMap::from([
-            (main.id.clone(), main),
-            (worktree_b.id.clone(), worktree_b),
-            (worktree_a.id.clone(), worktree_a),
-            (unrelated.id.clone(), unrelated),
-        ]));
-
-        let scope = catalog_workspace_scope(&workspaces, "main")
-            .await
-            .expect("resolve scope");
-
-        let ids: Vec<_> = scope.into_iter().map(|entry| entry.id).collect();
-        assert_eq!(ids, vec!["main", "worktree-a", "worktree-b"]);
-    }
-
-    #[tokio::test]
-    async fn catalog_workspace_scope_keeps_worktree_selection_isolated() {
-        let main = workspace_entry("main", "Main", "/tmp/main", WorkspaceKind::Main, None);
-        let worktree = workspace_entry(
-            "worktree-a",
-            "A",
-            "/tmp/worktree-a",
-            WorkspaceKind::Worktree,
-            Some("main"),
-        );
-        let sibling = workspace_entry(
-            "worktree-b",
-            "B",
-            "/tmp/worktree-b",
-            WorkspaceKind::Worktree,
-            Some("main"),
-        );
-        let workspaces = Mutex::new(HashMap::from([
-            (main.id.clone(), main),
-            (worktree.id.clone(), worktree),
-            (sibling.id.clone(), sibling),
-        ]));
-
-        let scope = catalog_workspace_scope(&workspaces, "worktree-a")
-            .await
-            .expect("resolve isolated scope");
-
-        let ids: Vec<_> = scope.into_iter().map(|entry| entry.id).collect();
-        assert_eq!(ids, vec!["worktree-a"]);
-    }
-
-    #[tokio::test]
-    async fn claude_child_workspace_session_is_not_claimed_by_parent_projection() {
-        let base = std::env::temp_dir().join(format!("claude-child-owner-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let repo_path = base.join("repo");
-        let child_path = repo_path.join("sub");
-        std::fs::create_dir_all(&child_path).expect("create child workspace path");
-
-        let claude_home = base.join("claude-home");
-        let claude_projects_dir = claude_home.join("projects");
-        let session_id = "child-claude-session";
-        write_claude_session_fixture(
-            &claude_projects_dir,
-            &repo_path,
-            session_id,
-            &child_path,
-            "child workspace task",
-        );
-
-        let parent = workspace_entry(
-            "parent",
-            "Parent",
-            &repo_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        let child = workspace_entry(
-            "child",
-            "Child",
-            &child_path.to_string_lossy(),
-            WorkspaceKind::Worktree,
-            Some("parent"),
-        );
-        let workspaces = Mutex::new(HashMap::from([
-            (parent.id.clone(), parent),
-            (child.id.clone(), child),
-        ]));
-        let engine_manager = engine::EngineManager::new();
-        engine_manager
-            .set_engine_config(
-                engine::EngineType::Claude,
-                engine::EngineConfig {
-                    home_dir: Some(claude_home.to_string_lossy().to_string()),
-                    ..engine::EngineConfig::default()
-                },
-            )
-            .await;
-
-        let parent_data = build_workspace_scope_catalog_data(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "parent",
-            SessionCatalogScanMode::Bounded(20),
-            WorkspaceSessionAttributionMode::Related,
-        )
-        .await
-        .expect("build parent catalog data");
-        let child_data = build_workspace_scope_catalog_data(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "child",
-            SessionCatalogScanMode::Bounded(20),
-            WorkspaceSessionAttributionMode::Related,
-        )
-        .await
-        .expect("build child catalog data");
-
-        let parent_claude_entries = parent_data
-            .entries
-            .iter()
-            .filter(|entry| entry.session_id == format!("claude:{session_id}"))
-            .collect::<Vec<_>>();
-        assert_eq!(parent_claude_entries.len(), 1);
-        assert_eq!(parent_claude_entries[0].workspace_id, "child");
-        assert_eq!(
-            parent_claude_entries[0].matched_workspace_id.as_deref(),
-            Some("child")
-        );
-
-        let child_claude_entries = child_data
-            .entries
-            .iter()
-            .filter(|entry| entry.session_id == format!("claude:{session_id}"))
-            .collect::<Vec<_>>();
-        assert_eq!(child_claude_entries.len(), 1);
-        assert_eq!(child_claude_entries[0].workspace_id, "child");
-        assert_eq!(
-            child_claude_entries[0].matched_workspace_id.as_deref(),
-            Some("child")
-        );
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_only_excludes_unrelated_claude_project_dir_with_matching_cwd() {
-        let base = std::env::temp_dir().join(format!("claude-workspace-only-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let workspace_path = base.join("workspace");
-        let unrelated_path = base.join("unrelated");
-        std::fs::create_dir_all(&workspace_path).expect("create workspace path");
-        std::fs::create_dir_all(&unrelated_path).expect("create unrelated path");
-
-        let claude_home = base.join("claude-home");
-        let claude_projects_dir = claude_home.join("projects");
-        write_claude_session_fixture(
-            &claude_projects_dir,
-            &unrelated_path,
-            "foreign-project-matching-cwd",
-            &workspace_path,
-            "foreign project but matching cwd",
-        );
-
-        let workspace = workspace_entry(
-            "workspace",
-            "Workspace",
-            &workspace_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-        engine_manager
-            .set_engine_config(
-                engine::EngineType::Claude,
-                engine::EngineConfig {
-                    home_dir: Some(claude_home.to_string_lossy().to_string()),
-                    ..engine::EngineConfig::default()
-                },
-            )
-            .await;
-
-        let related_page = list_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "workspace".to_string(),
-            Some(WorkspaceSessionCatalogQuery {
-                engine: Some("claude".to_string()),
-                status: Some("all".to_string()),
-                ..Default::default()
-            }),
-            None,
-            Some(20),
-        )
-        .await
-        .expect("list related sessions");
-        assert!(related_page
-            .data
-            .iter()
-            .any(|entry| entry.session_id == "claude:foreign-project-matching-cwd"));
-
-        let workspace_only_page = list_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "workspace".to_string(),
-            Some(WorkspaceSessionCatalogQuery {
-                engine: Some("claude".to_string()),
-                status: Some("all".to_string()),
-                session_attribution_mode: Some(WorkspaceSessionAttributionMode::WorkspaceOnly),
-                ..Default::default()
-            }),
-            None,
-            Some(20),
-        )
-        .await
-        .expect("list workspace-only sessions");
-
-        assert!(workspace_only_page
-            .data
-            .iter()
-            .all(|entry| entry.session_id != "claude:foreign-project-matching-cwd"));
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn workspace_only_keeps_claude_child_prefix_project_dir() {
-        let base = std::env::temp_dir().join(format!("claude-child-prefix-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let workspace_path = base.join("workspace");
-        let child_path = workspace_path.join("nested");
-        std::fs::create_dir_all(&child_path).expect("create child workspace path");
-
-        let claude_home = base.join("claude-home");
-        let claude_projects_dir = claude_home.join("projects");
-        write_claude_session_fixture(
-            &claude_projects_dir,
-            &child_path,
-            "child-prefix-session",
-            &child_path,
-            "child prefix task",
-        );
-
-        let workspace = workspace_entry(
-            "workspace",
-            "Workspace",
-            &workspace_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        let workspaces = Mutex::new(HashMap::from([(workspace.id.clone(), workspace)]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-        engine_manager
-            .set_engine_config(
-                engine::EngineType::Claude,
-                engine::EngineConfig {
-                    home_dir: Some(claude_home.to_string_lossy().to_string()),
-                    ..engine::EngineConfig::default()
-                },
-            )
-            .await;
-
-        let page = list_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "workspace".to_string(),
-            Some(WorkspaceSessionCatalogQuery {
-                engine: Some("claude".to_string()),
-                status: Some("all".to_string()),
-                session_attribution_mode: Some(WorkspaceSessionAttributionMode::WorkspaceOnly),
-                ..Default::default()
-            }),
-            None,
-            Some(20),
-        )
-        .await
-        .expect("list workspace-only sessions");
-
-        assert!(page
-            .data
-            .iter()
-            .any(|entry| entry.session_id == "claude:child-prefix-session"));
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn project_aggregate_child_claude_mutations_use_owner_stable_key() {
-        let base = std::env::temp_dir().join(format!("claude-child-mutate-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let repo_path = base.join("repo");
-        let child_path = repo_path.join("sub");
-        std::fs::create_dir_all(&child_path).expect("create child workspace path");
-
-        let claude_home = base.join("claude-home");
-        let claude_projects_dir = claude_home.join("projects");
-        let session_id = "child-mutation-session";
-        write_claude_session_fixture(
-            &claude_projects_dir,
-            &child_path,
-            session_id,
-            &child_path,
-            "child mutation task",
-        );
-        let child_project_dir = create_claude_project_dir(&claude_projects_dir, &child_path);
-        let session_path = child_project_dir.join(format!("{session_id}.jsonl"));
-
-        let parent = workspace_entry(
-            "parent",
-            "Parent",
-            &repo_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        let child = workspace_entry(
-            "child",
-            "Child",
-            &child_path.to_string_lossy(),
-            WorkspaceKind::Worktree,
-            Some("parent"),
-        );
-        let workspaces = Mutex::new(HashMap::from([
-            (parent.id.clone(), parent),
-            (child.id.clone(), child),
-        ]));
-        let sessions = Mutex::new(HashMap::new());
-        let engine_manager = engine::EngineManager::new();
-        engine_manager
-            .set_engine_config(
-                engine::EngineType::Claude,
-                engine::EngineConfig {
-                    home_dir: Some(claude_home.to_string_lossy().to_string()),
-                    ..engine::EngineConfig::default()
-                },
-            )
-            .await;
-
-        let requested_session_id = format!("claude:{session_id}");
-        let stable_key = format!("claude:child:{session_id}");
-        let archive_response = archive_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "parent".to_string(),
-            vec![requested_session_id.clone()],
-        )
-        .await
-        .expect("archive child from parent aggregate");
-
-        assert_eq!(archive_response.results.len(), 1);
-        assert!(archive_response.results[0].ok);
-        assert_eq!(
-            archive_response.results[0].owner_workspace_id.as_deref(),
-            Some("child")
-        );
-        assert_eq!(
-            archive_response.results[0].stable_session_key.as_deref(),
-            Some(stable_key.as_str())
-        );
-        let parent_metadata = read_catalog_metadata(&storage_path, "parent").expect("parent metadata");
-        let child_metadata = read_catalog_metadata(&storage_path, "child").expect("child metadata");
-        assert!(!parent_metadata
-            .archived_at_by_session_id
-            .contains_key(&stable_key));
-        assert!(child_metadata
-            .archived_at_by_session_id
-            .contains_key(&stable_key));
-
-        let unarchive_response = unarchive_workspace_sessions_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "parent".to_string(),
-            vec![requested_session_id.clone()],
-        )
-        .await
-        .expect("unarchive child from parent aggregate");
-
-        assert!(unarchive_response.results[0].ok);
-        assert_eq!(
-            unarchive_response.results[0].owner_workspace_id.as_deref(),
-            Some("child")
-        );
-        let child_metadata = read_catalog_metadata(&storage_path, "child").expect("child metadata");
-        assert!(!child_metadata
-            .archived_at_by_session_id
-            .contains_key(&stable_key));
-
-        let folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "child".to_string(),
-            "Child folder".to_string(),
-            None,
-        )
-        .await
-        .expect("create child folder")
-        .folder;
-        let move_response = assign_workspace_session_folders_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "parent".to_string(),
-            vec![requested_session_id.clone()],
-            Some(folder.id.clone()),
-        )
-        .await
-        .expect("move child row from parent aggregate");
-
-        assert!(move_response.results[0].ok);
-        assert_eq!(
-            move_response.results[0].owner_workspace_id.as_deref(),
-            Some("child")
-        );
-        let parent_metadata = read_catalog_metadata(&storage_path, "parent").expect("parent metadata");
-        let child_metadata = read_catalog_metadata(&storage_path, "child").expect("child metadata");
-        assert!(!parent_metadata
-            .folder_id_by_session_id
-            .contains_key(&stable_key));
-        assert_eq!(
-            child_metadata.folder_id_by_session_id.get(&stable_key),
-            Some(&folder.id)
-        );
-
-        let delete_response = delete_workspace_sessions_core(
-            &workspaces,
-            &sessions,
-            &engine_manager,
-            &storage_path,
-            "parent".to_string(),
-            vec![requested_session_id],
-        )
-        .await
-        .expect("delete child from parent aggregate");
-
-        assert!(delete_response.results[0].ok);
-        assert_eq!(
-            delete_response.results[0].owner_workspace_id.as_deref(),
-            Some("child")
-        );
-        assert!(!session_path.exists());
-        let child_metadata = read_catalog_metadata(&storage_path, "child").expect("child metadata");
-        assert!(!child_metadata
-            .folder_id_by_session_id
-            .contains_key(&stable_key));
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn folder_assignment_returns_partial_results_when_owner_folder_is_missing() {
-        let base = std::env::temp_dir().join(format!("folder-partial-owner-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let repo_path = base.join("repo");
-        let child_path = repo_path.join("sub");
-        std::fs::create_dir_all(&child_path).expect("create child workspace path");
-
-        let claude_home = base.join("claude-home");
-        let claude_projects_dir = claude_home.join("projects");
-        write_claude_session_fixture(
-            &claude_projects_dir,
-            &repo_path,
-            "parent-session",
-            &repo_path,
-            "parent mutation task",
-        );
-        write_claude_session_fixture(
-            &claude_projects_dir,
-            &child_path,
-            "child-session",
-            &child_path,
-            "child mutation task",
-        );
-
-        let parent = workspace_entry(
-            "parent",
-            "Parent",
-            &repo_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        let child = workspace_entry(
-            "child",
-            "Child",
-            &child_path.to_string_lossy(),
-            WorkspaceKind::Worktree,
-            Some("parent"),
-        );
-        let workspaces = Mutex::new(HashMap::from([
-            (parent.id.clone(), parent),
-            (child.id.clone(), child),
-        ]));
-        let engine_manager = engine::EngineManager::new();
-        engine_manager
-            .set_engine_config(
-                engine::EngineType::Claude,
-                engine::EngineConfig {
-                    home_dir: Some(claude_home.to_string_lossy().to_string()),
-                    ..engine::EngineConfig::default()
-                },
-            )
-            .await;
-
-        let parent_folder = create_workspace_session_folder_core(
-            &workspaces,
-            &storage_path,
-            "parent".to_string(),
-            "Parent folder".to_string(),
-            None,
-        )
-        .await
-        .expect("create parent folder")
-        .folder;
-
-        let response = assign_workspace_session_folders_core(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "parent".to_string(),
-            vec![
-                "claude:parent-session".to_string(),
-                "claude:child-session".to_string(),
-            ],
-            Some(parent_folder.id.clone()),
-        )
-        .await
-        .expect("partial folder assignment should stay request-successful");
-
-        assert_eq!(response.results.len(), 2);
-        let parent_result = response
-            .results
-            .iter()
-            .find(|result| result.session_id == "claude:parent-session")
-            .expect("parent result");
-        let child_result = response
-            .results
-            .iter()
-            .find(|result| result.session_id == "claude:child-session")
-            .expect("child result");
-        assert!(parent_result.ok);
-        assert_eq!(parent_result.owner_workspace_id.as_deref(), Some("parent"));
-        assert!(!child_result.ok);
-        assert_eq!(child_result.owner_workspace_id.as_deref(), Some("child"));
-        assert_eq!(
-            child_result.code.as_deref(),
-            Some("FOLDER_METADATA_UNAVAILABLE")
-        );
-
-        let parent_metadata = read_catalog_metadata(&storage_path, "parent").expect("parent metadata");
-        let child_metadata = read_catalog_metadata(&storage_path, "child").expect("child metadata");
-        assert_eq!(
-            parent_metadata
-                .folder_id_by_session_id
-                .get("claude:parent:parent-session"),
-            Some(&parent_folder.id)
-        );
-        assert!(child_metadata.folder_id_by_session_id.is_empty());
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[tokio::test]
-    async fn claude_independent_nested_workspace_session_is_not_claimed_by_parent_projection() {
-        let base = std::env::temp_dir().join(format!("claude-nested-owner-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&base).expect("create temp dir");
-        let storage_path = base.join("workspaces.json");
-        std::fs::write(&storage_path, "[]").expect("seed storage path");
-
-        let repo_path = base.join("repo");
-        let child_path = repo_path.join("sub");
-        std::fs::create_dir_all(&child_path).expect("create child workspace path");
-
-        let claude_home = base.join("claude-home");
-        let claude_projects_dir = claude_home.join("projects");
-        let session_id = "nested-claude-session";
-        write_claude_session_fixture(
-            &claude_projects_dir,
-            &repo_path,
-            session_id,
-            &child_path,
-            "nested workspace task",
-        );
-
-        let parent = workspace_entry(
-            "parent",
-            "Parent",
-            &repo_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        let child = workspace_entry(
-            "child",
-            "Child",
-            &child_path.to_string_lossy(),
-            WorkspaceKind::Main,
-            None,
-        );
-        let workspaces = Mutex::new(HashMap::from([
-            (parent.id.clone(), parent),
-            (child.id.clone(), child),
-        ]));
-        let engine_manager = engine::EngineManager::new();
-        engine_manager
-            .set_engine_config(
-                engine::EngineType::Claude,
-                engine::EngineConfig {
-                    home_dir: Some(claude_home.to_string_lossy().to_string()),
-                    ..engine::EngineConfig::default()
-                },
-            )
-            .await;
-
-        let parent_data = build_workspace_scope_catalog_data(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "parent",
-            SessionCatalogScanMode::Bounded(20),
-            WorkspaceSessionAttributionMode::Related,
-        )
-        .await
-        .expect("build parent catalog data");
-        let child_data = build_workspace_scope_catalog_data(
-            &workspaces,
-            &engine_manager,
-            &storage_path,
-            "child",
-            SessionCatalogScanMode::Bounded(20),
-            WorkspaceSessionAttributionMode::Related,
-        )
-        .await
-        .expect("build child catalog data");
-
-        assert!(!parent_data
-            .entries
-            .iter()
-            .any(|entry| entry.session_id == format!("claude:{session_id}")));
-
-        let child_claude_entry = child_data
-            .entries
-            .iter()
-            .find(|entry| entry.session_id == format!("claude:{session_id}"))
-            .expect("child projection should include nested claude session");
-        assert_eq!(child_claude_entry.workspace_id, "child");
-        assert_eq!(
-            child_claude_entry.matched_workspace_id.as_deref(),
-            Some("child")
-        );
-
-        std::fs::remove_dir_all(base).ok();
-    }
-
-    #[test]
-    fn catalog_entry_dedupe_key_includes_workspace_identity() {
-        let left = catalog_entry("shared-session", "main", None, None);
-        let right = WorkspaceSessionCatalogEntry {
-            workspace_id: "worktree-a".to_string(),
-            ..left.clone()
-        };
-
-        assert_ne!(
-            build_catalog_entry_dedupe_key(&left),
-            build_catalog_entry_dedupe_key(&right)
-        );
-    }
-
-    #[test]
-    fn catalog_entry_dedupe_key_collapses_same_workspace_identity() {
-        let left = catalog_entry("shared-session", "main", None, None);
-        let right = WorkspaceSessionCatalogEntry {
-            source: Some("override".to_string()),
-            source_label: Some("override/codex".to_string()),
-            updated_at: 2,
-            ..left.clone()
-        };
-
-        assert_eq!(
-            build_catalog_entry_dedupe_key(&left),
-            build_catalog_entry_dedupe_key(&right)
-        );
-    }
-
-    #[test]
-    fn partial_source_join_dedupes_scope_failures_without_dropping_signal() {
-        let partial_source = join_partial_sources(vec![
-            SESSION_CATALOG_PARTIAL_CODEX.to_string(),
-            SESSION_CATALOG_PARTIAL_GEMINI.to_string(),
-            SESSION_CATALOG_PARTIAL_CODEX.to_string(),
-        ]);
-
-        assert_eq!(
-            partial_source,
-            Some("codex-history-unavailable,gemini-history-unavailable".to_string())
-        );
-    }
-
-    #[test]
-    fn projection_summary_counts_filtered_total_separately_from_status_buckets() {
-        let mut active = catalog_entry("codex:active", "main", Some("Main"), None);
-        active.engine = "codex".to_string();
-        active.title = "Bugfix discussion".to_string();
-
-        let mut archived = catalog_entry("claude:archived", "worktree-a", Some("Worktree"), None);
-        archived.engine = "claude".to_string();
-        archived.title = "Bugfix archive".to_string();
-        archived.archived_at = Some(42);
-
-        let mut other = catalog_entry("gemini:other", "main", Some("Main"), None);
-        other.engine = "gemini".to_string();
-        other.title = "Other topic".to_string();
-
-        let counts = build_catalog_count_summary(
-            &[active, archived, other],
-            &WorkspaceSessionCatalogQuery {
-                keyword: Some("bugfix".to_string()),
-                engine: None,
-                status: Some("active".to_string()),
-                folder_id: None,
-                ..Default::default()
-            },
-        );
-
-        assert_eq!(
-            counts,
-            SessionCatalogCountSummary {
-                active_total: 1,
-                archived_total: 1,
-                all_total: 2,
-                filtered_total: 1,
-            }
-        );
-    }
-
-    #[test]
-    fn folder_count_summary_uses_filtered_entries_and_parent_folder_inheritance() {
-        let mut parent = catalog_entry("codex:parent", "main", Some("Main"), None);
-        parent.folder_id = Some("folder-a".to_string());
-        parent.title = "Bugfix parent".to_string();
-
-        let mut inherited_child = catalog_entry("codex:child", "main", Some("Main"), None);
-        inherited_child.parent_session_id = Some("codex:parent".to_string());
-        inherited_child.title = "Bugfix child".to_string();
-
-        let mut root = catalog_entry("codex:root", "main", Some("Main"), None);
-        root.title = "Bugfix root".to_string();
-
-        let mut filtered_out = catalog_entry("codex:other", "main", Some("Main"), None);
-        filtered_out.folder_id = Some("folder-a".to_string());
-        filtered_out.title = "Other topic".to_string();
-
-        let query = WorkspaceSessionCatalogQuery {
-            keyword: Some("bugfix".to_string()),
-            engine: None,
-            status: Some("active".to_string()),
-            folder_id: None,
-            ..Default::default()
-        };
-        let entries = [parent, inherited_child, root, filtered_out];
-        let filtered_entries = entries
-            .iter()
-            .filter(|entry| entry_matches_query(entry, &query))
-            .collect::<Vec<_>>();
-        let folder_counts = build_catalog_folder_count_summary(&filtered_entries);
-
-        assert_eq!(folder_counts.folder_counts_by_id.get("folder-a"), Some(&2));
-        assert_eq!(folder_counts.unassigned_folder_count, 1);
-    }
-
-    #[test]
-    fn catalog_page_filters_by_effective_folder_before_pagination() {
-        let mut newest = catalog_entry("codex:newest", "main", Some("Main"), None);
-        newest.updated_at = 300;
-
-        let mut parent = catalog_entry("codex:parent", "main", Some("Main"), None);
-        parent.updated_at = 200;
-        parent.folder_id = Some("folder-a".to_string());
-
-        let mut child = catalog_entry("codex:child", "main", Some("Main"), None);
-        child.updated_at = 100;
-        child.parent_session_id = Some("codex:parent".to_string());
-
-        let page = build_catalog_page(
-            vec![newest, parent, child],
-            WorkspaceSessionCatalogQuery {
-                keyword: None,
-                engine: None,
-                status: Some("active".to_string()),
-                folder_id: Some("folder-a".to_string()),
-                ..Default::default()
-            },
-            None,
-            Some(1),
-            None,
-            Vec::new(),
-        );
-
-        assert_eq!(page.data.len(), 1);
-        assert_eq!(page.data[0].session_id, "codex:parent");
-        assert!(page
-            .next_cursor
-            .as_deref()
-            .is_some_and(|cursor| cursor.starts_with(SESSION_CATALOG_STABLE_CURSOR_PREFIX)));
-    }
-
-    #[test]
-    fn normalize_partial_sources_preserves_first_seen_order() {
-        let partial_sources = normalize_partial_sources(vec![
-            SESSION_CATALOG_PARTIAL_GEMINI.to_string(),
-            SESSION_CATALOG_PARTIAL_CODEX.to_string(),
-            SESSION_CATALOG_PARTIAL_GEMINI.to_string(),
-        ]);
-
-        assert_eq!(
-            partial_sources,
-            vec![
-                SESSION_CATALOG_PARTIAL_GEMINI.to_string(),
-                SESSION_CATALOG_PARTIAL_CODEX.to_string(),
-            ]
-        );
     }
